Invoke hierarchical_count_table module

Invoke hierarchical_count_table module

Usage

hierarchical_count_table_ui(id)

hierarchical_count_table_server(
  id,
  table_dataset,
  pop_dataset,
  subjid_var,
  show_modal_on_click = TRUE,
  on_sbj_click_fun = function() NULL,
  default_hierarchy = NULL,
  default_group = NULL,
  hierarchy_choices = NULL,
  group_choices = NULL,
  intended_use_label = NULL
)

mod_hierarchical_count_table(
  module_id,
  table_dataset_name,
  pop_dataset_name,
  subjid_var = "USUBJID",
  show_modal_on_click = TRUE,
  default_hierarchy = NULL,
  default_group = NULL,
  hierarchy_choices = NULL,
  group_choices = NULL,
  intended_use_label =
    "Use only for internal review and monitoring during the conduct of clinical trials.",
  receiver_id = NULL
)

Arguments

id

character(0) The ID for the event count module instance.

table_dataset

data.frame A reactive dataset containing the event data.

pop_dataset

data.frame A reactive dataset containing the population data.

subjid_var

character(1) A string representing the subject identifier column in both datasets.

show_modal_on_click

logical(1) A flag to indicate whether clicking a table cell should display a modal dialog with the subject IDs.

on_sbj_click_fun

'function()'

Function to invoke when a subject is clicked

default_hierarchy

character(1|2)|NULL A default value for the hierarchy variables (optional).

default_group

character(1)|NULL A default value for the group variable (optional).

hierarchy_choices

character(1+)|NULL A character vector specifying the possible choices for the hierarchy variables (optional).

group_choices

character(1+)|NULL A character vector specifying the possible choices for the group variable (optional).

intended_use_label

Either a string indicating the intended use for export, or NULL. The provided label will be displayed prior to the download and will also be included in the exported file.

module_id

[character(1)]

Module Shiny id

table_dataset_name, pop_dataset_name

[character(1)]

Name of the dataset

receiver_id

character(1)

Shiny ID of the module receiving the selected subject ID in the data listing. This ID must be present in the app or be NULL.

table_dataset_disp, pop_dataset_disp

[mm_dispatcher(1)]

Dataset dispatcher. This parameter is incompatible with its *_dataset_name counterpart. Only for advanced use.

Value

A shiny::tagList containing the user interface for selecting hierarchy, group, and minimum percentage for event counting.

A reactive value containing the list of subjects in the clicked cell, if applicable.

Functions

hierarchical_count_table_ui(): UI for the event count module

hierarchical_count_table_server(): server Server logic for the event count module